Added edit option

+++ src/app/components/assessment/assessment-card/assessment-card.component.ts
@@ -1,6 +1,8 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { Router } from '@angular/router';
 import Assessment from 'src/app/model/Assessment';
+import CreateAssessmentMode from 'src/app/model/enums/CreateAssessmentMode';
+import Tag from 'src/app/model/Tag';
 import { AssessmentService } from 'src/app/service/assessment/assessment.service';
 import Swal from 'sweetalert2';
 import { CreateAssessmentService } from '../create-assessment/create-assessment.service';
@@ -13,6 +15,8 @@ import { CreateAssessmentService } from '../create-assessment/create-assessment.
 export class AssessmentCardComponent implements OnInit {
   @Input() assessment: Assessment = {};
 
+  @Output() searchTag = new EventEmitter();
+
   constructor(
     private assessmentService: AssessmentService,
     private router: Router,
@@ -22,21 +26,30 @@ export class AssessmentCardComponent implements OnInit {
   ngOnInit(): void {}
 
   deleteAssessment(event: Event) {
-    this.assessmentService
-      .deleteAssessmentByIdFromServer(this.assessment.id)
-      .subscribe((data) => {
-        if (data.success) {
-          this.assessmentService.deleteAssessmentByIdFromList(
-            this.assessment.id
-          );
-        }
+    // confirm delete action
+    Swal.fire({
+      title: 'Do you really want to delete assessment ?',
+      showCancelButton: true,
+      confirmButtonText: `Confirm`,
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.assessmentService
+          .deleteAssessmentByIdFromServer(this.assessment.id)
+          .subscribe((data) => {
+            if (data.success) {
+              this.assessmentService.deleteAssessmentByIdFromList(
+                this.assessment.id
+              );
+            }
 
-        Swal.fire({
-          icon: data.success ? 'success' : 'error',
-          title: data.success ? 'Success' : 'Error',
-          text: data.message,
-        });
-      });
+            Swal.fire({
+              icon: data.success ? 'success' : 'error',
+              title: data.success ? 'Success' : 'Error',
+              text: data.message,
+            });
+          });
+      }
+    });
   }
 
   viewScores(event: Event) {
@@ -46,7 +59,14 @@ export class AssessmentCardComponent implements OnInit {
   }
 
   editAssessment(event: Event) {
-    this.createAssessmentService.setAssessment(this.assessment)
+    this.createAssessmentService.setAssessment(this.assessment);
+    this.createAssessmentService.setCreateAssessmentMode(
+      CreateAssessmentMode.EDIT
+    );
     this.router.navigate(['/app/assessment/create']);
+  }
+
+  searchByTag(event: Event, tag: Tag) {
+    this.searchTag.emit(tag);
   }
 }
